Refacto Tickets page
